refactor: move crumbs test to backend tests
